refactor: modularize routes and data service
